I finally got around to changing the double-buffer for the window from a Gtk+ Pixmap to a Cairo Surface. I introduced C preprocessor (CPP) flags that can be set to switch back and forth between using Cairo Surfaces and Gtk+ Pixmaps as the double-buffer, just in case Gtk+ Pixmaps turn out to be more the useful, feature-rich APIs for reasons that I cannot foresee. I am leaving these CPP flags in the source code because I am still not certain I want to commit to either implementation.
I also was experimenting with capturing events using a Gtk+
EventBox. I added in some CPP flags to be able to easily switch
between using an EventBox and using just the Window itself to capture
events. As far as I can tell, EventBoxes add no benefit to the
program, but I am leaving the code there and able to activate with C
preprocessor flags if, in the future, I discover some use for
EventBoxes.