Updated instance of Worker API for cooperative threading models.
The previous Worker API assumed all threading would be handled by the
Haskell runtime. However Gtk+ does not provide thread-safety for
window resources and so making use of workers lead to race conditions
and segfaults. This new API demands more of the back-end providers by
asking providers to fork threads in the way most suited to the
back-end.