Bug fix: deadlock on initial evaluation of worker thread.
This deadlock was a result of the following conditions:

1. Gtk+ does cooperative multitasking rather than actually launching
   an OS thread,

2. the semantics of the Happlets Worker API requires for threads
   launched with 'WorkCycleWait' to evaluate the worker task
   immediately,

3. the thread that launches the worker was (before this bug fix) the
   same thread that performs the initial evaluation of the task,

4. the 'guiWorker' function in the "Happlets.GUI" module evaluates the
   'inOtherThreadRunGUI' function on every evaluation of the Worker
   task,

5. the 'inOtherThreadRunGUI' function performs locking on the GUI
   MVars under the assumption it is running in a different thread,
   which (in the case of Gtk+) is not true, it is the same thread.

This results in the 'inOtherThreadRunGUI' function being evaluated
within the same thread that "spawned" the worker, which means the MVar
is locked twice in a row.

This bug fix simply executes a one-shot worker that performs the
initial evaluation of the Worker task with a wait time of zero
seconds, in this way the semantics of the 'WorkCycleWait' constructor
is honored.