Turned ConfigState to lens functions, split into their own classes
Several of the stateful functions were using 'ConfigState' because
this modeled how Cairo works. However this is not very consistent with
how the rest of the Happlets API, which use Lenses to pure data types
to set stateful information. I changed the Happlets Types2D API to be
more consistent and use Lenses rather than ConfigState variables.