Remove ALSA API provider from the Gtk+ API provider
The Haskell ALSA foreign bindings library is no longer up-to-date and
cannot be built with GHC version 8.8 or later. Furthermore, the ALSA
provider should be in it's own package, not in the GTK+ package. I may
be able to update the ALSA bindings upstream, but when I do I will
have a separate Happlets audio provider package, it will not be
included in this package.